Pokemon now appears 5% of the time i hope I didn't break anything didn't fully test everything lol. she'll be right.

+++ Assets/Scripts/Minigames/Pokemon/Attack.cs
@@ -10,8 +10,8 @@ public class Attack : ScriptableObject
     public string animationTrigger;
 
     //I've decided to do it like this instead of using enums or something
-    public int damage = -10000; //default is -10000 just in case we want something that functions like an attack but does 0 damage or a negative idk
-    public int healAmount = -10000; //if this is set to something other than -10000, it will heal the user for that amount
+    public float damage = -10000; //default is -10000 just in case we want something that functions like an attack but does 0 damage or a negative idk
+    public float healAmount = -10000; //if this is set to something other than -10000, it will heal the user for that amount
     public int accuracy = 100; //chance to hit, from 0 to 100
 
     public int AnimationReturner()
